So far so good, glad that I was able to hook things up and get them working with hooks at least with my Enzo pedal.
Need to start working on the MIDI IO with hooks so that I can
start adding/removing pedals and keep track of that with hooks.

Once that's all set up, then I can start saving the state to
local storage and then try to build the state back up from there.