Made some progress...
Immer has been really helpful with managing state.

I think I need to make some _slight_ refactorings to the pedal
and midi device maps. I think I need to add the initial states
there and make that the pattern for all future devices.

It might be worth renaming the components to be more generic
in terms of the midi devices...